+++ src/NetworkCore/Network/Pipeline/NetworkPipelineInstaller.cs
@@ -1,0 +1,62 @@
+﻿/*
+ * Sovereign Engine
+ * Copyright (c) 2019 opticfluorine
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a 
+ * copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation 
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the 
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+ * DEALINGS IN THE SOFTWARE.
+ */
+
+using Castle.MicroKernel.Registration;
+using Castle.MicroKernel.SubSystems.Configuration;
+using Castle.Windsor;
+using Sovereign.EngineUtil.IoC;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Sovereign.NetworkCore.Network.Pipeline
+{
+
+    /// <summary>
+    /// IoC installer for the network pipelines.
+    /// </summary>
+    public sealed class NetworkPipelineInstaller : IWindsorInstaller
+    {
+        public void Install(IWindsorContainer container, IConfigurationStore store)
+        {
+            container.Register(EngineClasses.EngineAssemblies()
+                .BasedOn<IInboundPipelineStage>()
+                .WithServiceDefaultInterfaces()
+                .LifestyleSingleton()
+                .AllowMultipleMatches());
+
+            container.Register(EngineClasses.EngineAssemblies()
+                .BasedOn<IOutboundPipelineStage>()
+                .WithServiceDefaultInterfaces()
+                .LifestyleSingleton()
+                .AllowMultipleMatches());
+
+            container.Register(Component.For<InboundNetworkPipeline>()
+                .LifestyleSingleton());
+
+            container.Register(Component.For<OutboundNetworkPipeline>()
+                .LifestyleSingleton());
+        }
+    }
+
+}